Rotation with Trim

+++ Rotation/rotation.c
@@ -92,7 +92,6 @@ SDL_Surface *trim(SDL_Surface *image)
             }
         }
     }
-    // printf("Bottom -> %zu\n",bottom_trim);
     for (size_t j = 0; j < h; j++)
     {
         for (size_t i = 0; i < w; i++)
@@ -106,7 +105,6 @@ SDL_Surface *trim(SDL_Surface *image)
             }
         }
     }
-    // printf("Top -> %zu\n",top_trim);
     for (size_t i = 0; i < w; i++)
     {
         for (size_t j = top_trim; j > bottom_trim; j--)
@@ -120,7 +118,6 @@ SDL_Surface *trim(SDL_Surface *image)
             }
         }
     }
-    // printf("Right -> %zu\n",right_trim);
     for (size_t i = w; i > 0; i--)
     {
         for (size_t j = top_trim; j > bottom_trim; j--)
@@ -134,7 +131,6 @@ SDL_Surface *trim(SDL_Surface *image)
             }
         }
     }
-    // printf("Left -> %zu\n",left_trim);
     SDL_Surface *output_image = SDL_CreateRGBSurface(0, w-(right_trim-left_trim), h - (top_trim-bottom_trim), 32, 0, 0, 0, 0);
     for (size_t i = left_trim; i < right_trim; i++)
     {
@@ -151,21 +147,12 @@ SDL_Surface *trim(SDL_Surface *image)
 int *new_sheer(double angle, int x, int y)
 {
     int *x_and_y = malloc(sizeof(int) * 2);
-    //sheer 1
     double intan = angle / 2.f;
-    // printf("intan -> %.6f\n",intan);
     double tangent = tan(intan);
-    // printf("tangent for angel -> %f: is -> %.6f\n",angle,tangent);
     int new_x = round((double)x - ((double)y * tangent));
-    // printf("new_x -> %d \n",new_x);
     int new_y = y;
-    // printf("new_y -> %d \n",new_y);
-    //sheer 2
     new_y = round((double)new_x * sin(angle) + (double)new_y);
-    // printf("new_y -> %d \n",new_y);
-    //sheer 3
     new_x = round((double)new_x - ((double)new_y * tangent));
-    // printf("new_x -> %d \n",new_x);
     x_and_y[0] = new_x;
     x_and_y[1] = new_y;
     return x_and_y;
@@ -181,14 +168,14 @@ SDL_Surface *rotate(SDL_Surface *image, double angle)
     double sinus = sin(angle);
     int height = image->h;
     int width = image->w;
-    int new_height = round((abs(height * cosine) + abs(width * sinus))) + 1;
-    int new_width = round((abs(width * cosine) + abs(height * sinus))) + 1;
+    int new_height = round(((double)abs(height * cosine) + (double)abs(width * sinus))) + 1;
+    int new_width = round(((double)abs(width * cosine) + (double)abs(height * sinus))) + 1;
     SDL_Surface *output_image = SDL_CreateRGBSurface(0, new_width * 2, new_height * 2, 32, 0, 0, 0, 0);
     int original_center_height = round(((height + 1) / 2) - 1);
     int original_center_width = round(((width + 1) / 2) - 1);
-    for (size_t i = 0; i < height; i++)
+    for (int i = 0; i < height; i++)
     {
-        for (size_t j = 0; j < width; j++)
+        for (int j = 0; j < width; j++)
         {
             int y = height - 1 - i - original_center_height;
             int x = width - 1 - j - original_center_width;
